modules, cleaning and first round of feedback
