+++ profile/change_profile.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="/dist/css/main.css" />
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
+      crossorigin="anonymous"
+    ></script>
+    <script type="module" src="/profile/change_profile.js"></script>
+    <title>Social Media Application</title>
+  </head>
+  <body class="bg-primary">
+    <nav class="nav navbar d-flex justify-content-center">
+      <a class="btn" href="/feed/index.html">Latest posts</a>
+      <a class="btn" aria-current="page" href="/profile/index.html"
+        >My profile</a
+      >
+    </nav>
+    <main>
+      <h1 class="d-flex justify-content-center">Update profile</h1>
+      <divc class="d-flex justify-content-center">
+        <form class="container p-3 col-8">
+          <div class="p-3">
+            <input class="form-control p-3 rounded-pill image" type="text" placeholder="Link to profile picture">
+            <input type="text" class="form-control p-3 mt-3 rounded-pill username" placeholder="Username" />
+            <textarea type="text" class="form-control p-4 mt-3 rounded-pill profiletext" placeholder="Profile text"></textarea>
+          </div>
+          <div
+          class="alert alert-danger rounded pill m-4 invisible"
+          role="alert"
+          id="showError"
+        >
+          <p id="errorMessage">Error! Profile is not updated.</p>
+        </div>
+          <div class="m-4">
+            <a type="button" class="btn btn-dark btn-outline-secondary p-2 m-3 rounded-pill d-flex justify-content-center update">Update profile</a>
+          </div>
+          <a type="button" class="btn m-3 d-flex justify-content-center" href="/profile"
+          >Back</a
+        >
+        </form>
+      </div>
+    </main>
+  </body>
+</html>
