+++ profile/index.html
@@ -10,6 +10,7 @@
       crossorigin="anonymous"
     ></script>
     <script type="module" src="/search/search.js"></script>
+    <script type="module" src="/profile/profile_details.js"></script>
     <title>My profile</title>
   </head>
   <body class="bg-primary">
@@ -25,12 +26,12 @@
     <main>
       <div class="d-flex justify-content-center">
         <div class="card bg-light rounded col-11 col-md-8 col-lg-6">
-          <button
+          <a
             type="button"
-            class="btn btn-dark btn-outline-secondary m-4 col-4 rounded-pill"
+            class="btn btn-dark btn-outline-secondary m-3 col-3 rounded-pill"
+            href="/profile/change_profile.html"
+            >Update profile</a
           >
-            Follow
-          </button>
           <div class="d-flex justify-content-center">
             <p class="p-2 m-2">
               Lorem, ipsum dolor sit amet consectetur adipisicing elit. Itaque,
@@ -40,8 +41,8 @@
             </p>
             <div class="m-2 col-4">
               <img
-                src="../images/sobhan-joodi-Wl_N-A_Wp5U-unsplash.jpg"
-                class="img-thumbnail rounded-circle bg-secondary border border-0"
+                src="/images/sobhan-joodi-Wl_N-A_Wp5U-unsplash.jpg"
+                class="img-thumbnail rounded-circle bg-secondary border border-0 avatar"
                 alt="Woman with black hair and dark sunglasses standing outside"
               />
             </div>
@@ -64,9 +65,7 @@
                 class="collapse collapse-horizontal"
                 id="collapseWidthExample"
               >
-                <div class="card card-body">
-                  A list of follows will be presented here.
-                </div>
+                <div class="card card-body follows"></div>
               </div>
             </div>
             <p>
@@ -81,9 +80,17 @@
                 Followers
               </button>
             </p>
+            <div>
+              <div
+                class="collapse collapse-horizontal"
+                id="collapseWidthExample"
+              >
+                <div class="card card-body followers"></div>
+              </div>
+            </div>
           </div>
           <div class="d-flex justify-content-between p-2">
-            <h1 class="m-3">Username</h1>
+            <h1 class="m-3 username">Username</h1>
             <a type="button" class="btn m-3" href="../index.html">Log out</a>
           </div>
         </div>
